% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forwardSamplingTools.R
\name{doSeriesOfSimulateStopLimitPrices}
\alias{doSeriesOfSimulateStopLimitPrices}
\title{doSeriesOfSimulateStopLimitPrices}
\usage{
doSeriesOfSimulateStopLimitPrices(longFswXtDf, limitRatioSeq = seq(0.01, 0.1,
  0.01), stopRatioSeq = seq(0.01, 0.1, 0.01), NonHitIdxValue = 1e+06)
}
\arguments{
\item{longFswXtDf}{a forward sampling xts with long format}

\item{limitRatioSeq}{series of limit return ratio}

\item{stopRatioSeq}{series of stop return ratio}

\item{NonHitIdxValue}{when the price not hit stop or limit, assign this value}
}
\description{
a function which can help you transform your forward sampling xts to data.frame
}
\examples{
Xt = getSymbols('2330.TW', auto.assign = F)
Xt \%>\% Pt2Rt \%>\% ForwardSlidingWindow(20) \%>\% FswXt2Df(index(Xt)[c(1,3,5,7,9)],longFormat=T) \%>\% simulateStopLimitPrices(0.06,0.06) \%>\% head
}

