% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forwardSamplingTools.R
\name{ForwardSlidingWindow}
\alias{ForwardSlidingWindow}
\title{ForwardSlidingWindow}
\usage{
ForwardSlidingWindow(accRt, win_size)
}
\description{
ForwardSlidingWindow
}
\examples{
Xt = getSymbols('2330.TW', auto.assign = F)
Xt \%>\% Pt2Rt \%>\% ForwardSlidingWindow(20) \%>\% head
}

