% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forwardSamplingTools.R
\name{FswXt2Df}
\alias{FswXt2Df}
\title{FswXt2Df}
\usage{
FswXt2Df(fswXt, filterDatetimes = NULL)
}
\description{
FswXt2Df
}
\examples{
Xt = getSymbols('2330.TW', auto.assign = F)
Xt \%>\% Pt2Rt \%>\% ForwardSlidingWindow(20) \%>\% FswXt2Df(index(Xt)[c(1,3,5,7,9)]) \%>\% head
}

